default dump and load now use streaming pickle to ensure that only a subset of the full data is ever loaded in memory at any one time

+++ xsorted/__init__.py
@@ -28,38 +28,45 @@ __author__ = __copyright__ = "Daniel Bradburn"
 __license__ = "MIT"
 
 
-@contextmanager
-def serializer():
+def _dump(partition):
     """
-    Default serializer factory which uses pickle to serialize partitions
+    Dump the given partition to an external source.
 
-    :return: tuple of dump, load, where:
+    The default implementation is to pickle the list of objects to a temporary file.
 
-                dump -- a callable taking one parameter which is the iterable to serialize, the
-                        function should return an id which can be used to reload the serialized
-                        iterable.
+    :param partition: The partition of objects to dump.
 
-                load -- a callable taking an id, the function should return an iterable or
-                        generator which iterates the serialized iterable. It is important that
-                        this function does not materialize the iterable.
+    :return: Unique id which can be used to reload the serialized partition. In the case of the
+             default implementation this is the path to the temporary file.
     """
-    partition_ids = []
+    with tempfile.NamedTemporaryFile(delete=False) as fileobj:
+        for item in partition:
+            pickle.dump(item, fileobj)
+        return fileobj.name
 
-    def dump(partition):
-        with tempfile.NamedTemporaryFile(delete=False) as fileobj:
-            pickle.dump(partition, fileobj)
-            partition_ids.append(fileobj.name)
-            return fileobj.name
 
-    def load(partition_id):
-        with open(partition_id, 'rb') as fileobj:
-            return pickle.load(fileobj)  # todo : need to use streaming
+def _load(partition_id):
+    """
+    Load a partition from an external source.
 
-    try:
-        yield dump, load
-    finally:
-        for partition_id in filter(os.path.exists, partition_ids):
+    The default implementation yields items loaded and unpickled from a temporary file. After all
+    items have been loaded the temporary file is removed.
+
+    :param partition_id: Unique identifier which can be used to reload the partition. In the case
+                         of the default implemenation this is the path to the temporary file to
+                         load.
+
+    :return: iterable which is loaded from the external source using partition_id.
+    """
+    if os.path.exists(partition_id):
+        try:
+            with open(partition_id, 'rb') as fileobj:
+                while fileobj.peek(1):
+                    yield pickle.load(fileobj)
+        finally:
             os.unlink(partition_id)
+    else:
+        raise StopIteration()
 
 
 def _split(dump, partition_size, iterable, key=None, reverse=False):
@@ -70,18 +77,22 @@ def _split(dump, partition_size, iterable, key=None, reverse=False):
     :param dump:            Callable which takes an iterable and serializes to some external
                             source, returning an iterable of ids which can be used to reload the
                             externalized partitions.
+
     :param partition_size:  The number of items to place in each partition.
+
     :param iterable:        The iterable to split into sorted partitions.
+
     :param key:             Callable which is used to retrieve the field to sort by.
+
     :param reverse:         If set to ``True``, then the list elements are sorted as if each
                             comparison were reversed.
 
     :return: iterable of the ids which can be used to reload the externalized partitions.
     """
-    sorted_bound = partial(sorted, key=key, reverse=reverse)
+    sort_by_key_and_maybe_reverse = partial(sorted, key=key, reverse=reverse)
     partitioned = partition_all(partition_size, iterable)
-    dump_sorted = compose(dump, sorted_bound)
-    return map(dump_sorted, partitioned)
+    dump_sorted = compose(dump, sort_by_key_and_maybe_reverse)
+    return list(map(dump_sorted, partitioned))
 
 
 def _merge(load, partition_ids, key=None, reverse=False):
@@ -90,8 +101,11 @@ def _merge(load, partition_ids, key=None, reverse=False):
 
     :param load:          Callable which loads and returns an iterable for iterating the
                           externalized partitions.
+
     :param partition_ids: Ids which can be used to reload the serialized partitions.
+
     :param key:           ``sorted`` key parameter.
+
     :param reverse:       ``sorted`` reverse parameter.
 
     :return: iterable of merged partitions.
@@ -99,18 +113,30 @@ def _merge(load, partition_ids, key=None, reverse=False):
     return merge(*map(load, partition_ids), key=key, reverse=reverse)
 
 
-def _xsorted(partition_size, serializer_factory, split, merge, iterable, key=None, reverse=False):
+def _xsorted(partition_size, dump, load, split, merge, iterable, key=None, reverse=False):
     """
     xsorted implementation where dependencies should be injected, athough it is possible to use
     this function directly the xsorter function should be used to pre-bind the dependencies for
     convenience.
 
     :param partition_size:     The number of items to serialize in each partition.
-    :param serializer_factory: Callable returning a context manager which returns a tuple of
-                               (dump, load) on ``__enter__``.
+
+    :param dump:               Callable taking one parameter which is the iterable to
+                               serialize, the function should return an id which can be
+                               used to reload the serialized iterable.
+
+    :param load:               Callable taking an id, the function should return an
+                               iterable or generator which iterates the serialized
+                               iterable. It is important that this function does not
+                               materialize the iterable, otherwise the whole data set will
+                               be loaded in memory, which defeats the object of this kind
+                               of sorting.
+
     :param split:              Callable which is used to sort and split the given iterable into
                                partitions.
+
     :param merge:              Callable which is used to merge and sort serialized partitions.
+
     :param iterable:           The iterable to be sorted.
 
     :param key:                Specifies a function of one argument that is used to extract a
@@ -122,38 +148,36 @@ def _xsorted(partition_size, serializer_factory, split, merge, iterable, key=Non
 
     :return: an iterable which returns the elements of the input iterable in sorted order.
     """
-    with serializer_factory() as (dump, load):
-        partition_ids = split(dump, partition_size, iterable, key, reverse)
-        return merge(load, partition_ids, key, reverse)
+    partition_ids = split(dump, partition_size, iterable, key, reverse)
+    return merge(load, partition_ids, key, reverse)
 
 
-def xsorter(partition_size=8192, serializer_factory=serializer, split=_split, merge=_merge):
+def xsorter(partition_size=8192, dump=_dump, load=_load, split=_split, merge=_merge):
     """
     Generate an xsorted function using the specified partition size, serializer factory, splitter
     and merger.
 
     :param partition_size:     The number of items to serialize in each partition.
-    :param serializer_factory: Context manager which should return a tuple dump and load for
-                               serializing data.
 
-                                dump -- a callable taking one parameter which is the iterable to
-                                        serialize, the function should return an id which can be
-                                        used to reload the serialized iterable.
+    :param dump:               Callable taking one parameter which is the iterable to
+                               serialize, the function should return an id which can be
+                               used to reload the serialized iterable.
 
-                                load -- a callable taking an id, the function should return an
-                                        iterable or generator which iterates the serialized
-                                        iterable. It is important that this function does not
-                                        materialize the iterable, otherwise the whole data set will
-                                        be loaded in memory, which defeats the object of this kind
-                                        of sorting.
+    :param load:               Callable taking an id, the function should return an
+                               iterable or generator which iterates the serialized
+                               iterable. It is important that this function does not
+                               materialize the iterable, otherwise the whole data set will
+                               be loaded in memory, which defeats the object of this kind
+                               of sorting.
 
     :param split:              Callable which is used to sort and split the given iterable into
                                partitions.
+
     :param merge:              Callable which is used to merge and sort serialized partitions.
 
     :return: xsorted function.
     """
-    return partial(_xsorted, partition_size, serializer_factory, split, merge)
+    return partial(_xsorted, partition_size, dump, load, split, merge)
 
 
 def xsorted(iterable, key=None, reverse=False):
